♻️ Refactor: 작성 시간 변환 함수 utils로 추출해 재사용
