+++ src/components/Post/PostCard.jsx
@@ -7,6 +7,7 @@ import authInstance from '../../api/instance/authInstance';
 import { ReactComponent as HeartIcon } from '../../assets/icons/icon-heart.svg';
 import { ReactComponent as CommentIcon } from '../../assets/icons/icon-message-small.svg';
 import { ReactComponent as MoreIcon } from '../../assets/icons/icon-more-small.svg';
+import { convertTime } from '../../utils/convertTime';
 
 const PostArticle = styled.article`
   position: relative;
@@ -121,13 +122,8 @@ export default function PostCard({ post, setPostId, setIsMenuOpen }) {
   const navigate = useNavigate();
   const pathname = useLocation().pathname;
   const isProfile = pathname.includes('profile');
-
-  const date = `
-    ${createdAt.slice(0, 4)}년 
-    ${createdAt.slice(5, 7)}월 
-    ${createdAt.slice(8, 10)}일
-  `;
   const imageList = image.split(',');
+
   const handleLike = async () => {
     try {
       const endpoint = likeInfo.liked
@@ -136,7 +132,6 @@ export default function PostCard({ post, setPostId, setIsMenuOpen }) {
       const res = await (likeInfo.liked
         ? authInstance.delete(endpoint)
         : authInstance.post(endpoint));
-      console.log(res);
       setLikeInfo({
         liked: res.data.post.hearted,
         count: res.data.post.heartCount,
@@ -207,7 +202,7 @@ export default function PostCard({ post, setPostId, setIsMenuOpen }) {
             <IconText>{commentCount}</IconText>
           </div>
         </InfoIcons>
-        <time dateTime={createdAt}>{date}</time>
+        <time dateTime={createdAt}>{convertTime(createdAt)}</time>
       </ContentInfo>
       <StyledIconButton>
         <MoreIcon onClick={handleMenu} />
