+++ src/utils/convertTime.js
@@ -1,0 +1,28 @@
+export const convertTime = (date) => {
+  const currentTime = new Date();
+  const commentDate = new Date(date);
+  const timeDifference = currentTime - commentDate;
+
+  const seconds = Math.floor(timeDifference / 1000);
+  const minutes = Math.floor(seconds / 60);
+  const hours = Math.floor(minutes / 60);
+  const days = Math.floor(hours / 24);
+  const weeks = Math.floor(days / 7);
+  const years = Math.floor(days / 365);
+
+  if (seconds === 0) {
+    return '방금';
+  } else if (seconds < 60) {
+    return `${seconds}초 전`;
+  } else if (minutes < 60) {
+    return `${minutes}분 전`;
+  } else if (hours < 24) {
+    return `${hours}시간 전`;
+  } else if (days < 7) {
+    return `${days}일 전`;
+  } else if (weeks < 52) {
+    return `${weeks}주 전`;
+  } else {
+    return `${years}년 전`;
+  }
+};
